requirements.txt
ml-collections==0.1.0
pandas==1.2.3
Pillow==8.1.2
torch==1.8.0
torchvision==0.9.0
numpy==1.20.1
scipy==1.6.2
tqdm==4.60.0